requirements.txt
numpy
streamlit
pandas
keras
Keras-Preprocessing
tensorflow-cpu
matplotlib
pillow
opencv-python